+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.3] - 2026-02-05
+- fix #5, improve demo sketch
+- fix sign extension in **readCurrentOffset()**
+- add DS2438_time_convert.ino
+- update readme.md
+
+
 ## [0.1.2] - 2026-01-02
 - update GitHub actions
 - update examples